Re-structure code
