+++ batchflow/models/torch/__init__.py
@@ -12,3 +12,5 @@ from .vnet import VNet
 from .densenet import SegmentationDenseNet, DenseNetFC56, DenseNetFC67, DenseNetFC103
 from .utils import get_shape, get_num_channels, get_num_dims, calc_padding, unpack_fn_from_config, safe_eval
 from .blocks import DefaultBlock, XceptionBlock, VGGBlock, ResBlock, DenseBlock
+
+from .lightning import LightningModel, PipelineDataset, PipelineDataLoader
